+++ src/store/store.cache.js
@@ -6,6 +6,34 @@
  * ]
  */
 
+// https://stackoverflow.com/questions/7837456/how-to-compare-arrays-in-javascript/14853974
+Array.prototype.equals = function (array) {
+  // if the other array is a falsy value, return
+  if (!array) {
+    return false;
+  }
+
+  // compare lengths - can save a lot of time
+  if (this.length !== array.length) {
+    return false;
+  }
+
+  for (var i = 0, l=this.length; i < l; i++) {
+    // Check if we have nested arrays
+    if (this[i] instanceof Array && array[i] instanceof Array) {
+      // recurse into the nested arrays
+      if (!this[i].equals(array[i])) {
+        return false;
+      }
+    }
+    else if (this[i] !== array[i]) {
+      // Warning - two different object instances will never be equal: {x:20} != {x:20}
+      return false;
+    }
+  }
+  return true;
+};
+
 /**
  * Cache object
  * @return {Object}
@@ -41,6 +69,9 @@ function cache () {
           break;
         }
         if (_filter === filterValues[_filterValueKeys[k]]) {
+          _isFilterMatchValue = true;
+        }
+        else if (Array.isArray(_filter) && _filter.equals(filterValues[_filterValueKeys[k]])) {
           _isFilterMatchValue = true;
         }
         else {